CS-5833: Move logic out of the init file, and add a link pointing to the projects in the users CS instance
